etl/pyproject.toml
[project]
name = "echerga-stats"
version = "0.1.0"
description = "Echerga border crossing statistics collector"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
]
